[BACKLOG-12260]  Support for steps like Lookup.
Rows coming into lookup don't necessarily correspond to
a processRow call, since the processRow call will block
waiting for all of the lookup input.
Now running the processRow loop in a separate thread.

+++ pdi-execution-engine/api/src/main/java/org/pentaho/di/engine/api/IProgressReporting.java
@@ -18,21 +18,21 @@ public interface IProgressReporting<T extends IPDIEvent> extends IPDIEventSource
    *
    * @return
    */
-  int getIn();
+  long getIn();
 
   /**
    * Get number of {@link IPDIEvent}s out from this component
    *
    * @return
    */
-  int getOut();
+  long getOut();
 
   /**
    * Get number of {@link IPDIEvent}s dropped (errorred)
    *
    * @return
    */
-  int getDropped();
+  long getDropped();
 
   /**
    * Get number of {@link IPDIEvent}s currently in-flight
